+++ libraries/exoplayer_rtsp/src/main/java/androidx/media3/exoplayer/rtsp/RtpPacket.java
@@ -205,6 +205,7 @@ public final class RtpPacket {
     byte version = (byte) (firstByte >> 6);
     boolean padding = ((firstByte >> 5) & 0x1) == 1;
     byte csrcCount = (byte) (firstByte & 0xF);
+    boolean hasExtension = ((firstByte >> 4) & 0x1) == 1;
 
     if (version != RTP_VERSION) {
       return null;
@@ -231,6 +232,16 @@ public final class RtpPacket {
       }
     } else {
       csrc = EMPTY;
+    }
+
+    // Extension.
+    if (hasExtension) {
+      // Skip profile-defined data
+      packetBuffer.skipBytes(2);
+      int headerExtensionPayloadLength = packetBuffer.readShort();
+      if (headerExtensionPayloadLength != 0) {
+        packetBuffer.skipBytes(headerExtensionPayloadLength * 4);
+      }
     }
 
     // Everything else will be RTP payload.