+++ RELEASENOTES.md
@@ -47,6 +47,9 @@
         per sample.
     *   Fix playback hanging on DASH multi-period streams when CEA-608 subtitles
         are enabled ([#1863](https://github.com/androidx/media/issues/1863)).
+    *   Fix garbled CEA-608 subtitles in MP4 files that incorrectly mark every
+        sample as a sync sample
+        ([#1863](https://github.com/androidx/media/issues/1863)).
 *   Metadata:
 *   Image:
 *   DRM: